telegram-bot/requirements.txt
python-telegram-bot>=22.0
google-genai>=1.0
aiohttp
Pillow
